odict review
